+++ src/main/scala/org/podval/tools/scalajsplugin/scalajs/ScalaJSTestTask.scala
@@ -1,6 +1,7 @@
 package org.podval.tools.scalajsplugin.scalajs
 
 import org.podval.tools.build.ScalaBackendKind
+import org.podval.tools.scalajs.ScalaJSTestEnvironment
 import org.podval.tools.scalajsplugin.nonjvm.NonJvmRunTask
 
 abstract class ScalaJSTestTask extends NonJvmRunTask.Test[ScalaJSLinkTask] with ScalaJSRunTask:
@@ -8,6 +9,9 @@ abstract class ScalaJSTestTask extends NonJvmRunTask.Test[ScalaJSLinkTask] with 
 
   final override protected def backendKind: ScalaBackendKind = ScalaBackendKind.JS
 
-  final override protected def createTestEnvironment: ScalaJSTestEnvironment = ScalaJSTestEnvironment(
-    scalaJSRunCommon
-  )
+  final override protected def createTestEnvironment: ScalaJSTestEnvironment =
+    scalaJSRun.createTestEnvironment(
+      jsDirectory = linkTask.getJSDirectory,
+      node = linkTask.node,
+      logSource = getName
+    )