+++ src/main/scala/org/podval/tools/scalajsplugin/scalanative/ScalaNativeRunTask.scala
@@ -1,7 +1,6 @@
 package org.podval.tools.scalajsplugin.scalanative
 
 import org.podval.tools.node.Node
-import org.podval.tools.scalajs.ScalaJSRunCommon
 import org.podval.tools.scalajsplugin.nonjvm.NonJvmRunTask
 
 trait ScalaNativeRunTask extends NonJvmRunTask[ScalaNativeLinkTask] with ScalaNativeTask