+++ src/test/scala/org/podval/tools/test/testproject/TestResultsRetriever.scala
@@ -3,15 +3,13 @@ package org.podval.tools.test.testproject
 import org.gradle.api.Action
 import org.gradle.api.internal.tasks.testing.junit.result.{TestClassResult, TestMethodResult, TestResultSerializer}
 import org.podval.tools.build.ScalaBackendKind
-import org.podval.tools.scalajsplugin.GradleNames
 import org.podval.tools.util.Files
 import scala.jdk.CollectionConverters.ListHasAsScala
 import java.io.File
 
 final class TestResultsRetriever(projectDir: File):
   def testResults(backend: Option[ScalaBackendKind]): List[TestClassResult] =
-    val directory: String = backend.map(GradleNames.testSuiteName).getOrElse("test")
-    readTestClassResults(directory)
+    readTestClassResults(ScalaBackendKind.testSuiteName(backend))
 
   private def readTestClassResults(name: String): List[TestClassResult] =
     val binaryTestReportDir: File = Files.file(projectDir, "build", "test-results", name, "binary")